Minor edits. - minor addition to ui. - base work to eventually add more systems.
