+++ module.json
@@ -1,7 +1,7 @@
 {
   "id": "backpack-manager",
   "title": "Backpack Manager",
-  "version": "10.0.3",
+  "version": "10.0.4",
   "authors": [
     {
       "name": "Zhell",
@@ -13,7 +13,7 @@
   "compatibility": {
     "minimum": "10",
     "maximum": "10",
-    "verified": "10.286"
+    "verified": "10.287"
   },
   "esmodules": [
     "setup.js"
@@ -42,7 +42,7 @@
   "url": "https://github.com/krbz999/backpack-manager",
   "license": "https://raw.githubusercontent.com/krbz999/backpack-manager/main/LICENSE",
   "manifest": "https://github.com/krbz999/backpack-manager/releases/latest/download/module.json",
-  "download": "https://github.com/krbz999/backpack-manager/releases/download/v10.0.3/module.zip",
+  "download": "https://github.com/krbz999/backpack-manager/releases/download/v10.0.4/module.zip",
   "description": "A helpful UI to manage external actors that act as backpacks.",
   "name": "backpack-manager",
   "minimumCoreVersion": "10"
