+++ templates/backpack-manager.css
@@ -6,8 +6,10 @@ form.backpack-manager header {
   display: grid;
   align-content: space-evenly;
   font-size: 16px;
-  padding: 0 0 0 0.5em;
+  padding: 0 0.5em;
   border-radius: 3px;
+  justify-content: space-between;
+  grid-template-columns: auto auto;
 }
 
 form.backpack-manager .stowed .stowed-content,
@@ -30,6 +32,7 @@ form.backpack-manager .inventory-content .item {
   font-size: 1.1em;
   gap: 1em;
 }
+
 form.backpack-manager .stowed-content .item:nth-child(2n-1),
 form.backpack-manager .inventory-content .item:nth-child(2n-1) {
   background: #11643b21;
